+++ src/test/java/seedu/address/model/person/DateAddedIsBeforeDateInputPredicateTest.java
@@ -2,7 +2,7 @@ package seedu.address.model.person;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import static seedu.address.testutil.TypicalPersons.AMY;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.AMY;
 
 import org.junit.Test;
