+++ src/test/java/seedu/address/storage/XmlAddressBookStorageTest.java
@@ -2,10 +2,10 @@ package seedu.address.storage;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static seedu.address.testutil.TypicalPersons.ALICE;
-import static seedu.address.testutil.TypicalPersons.HOON;
-import static seedu.address.testutil.TypicalPersons.IDA;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.ALICE;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.HOON;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.IDA;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.getTypicalAddressBook;
 
 import java.io.IOException;
