+++ src/test/java/systemtests/AddressBookSystemTest.java
@@ -32,7 +32,7 @@ import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.commands.SelectCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
-import seedu.address.testutil.TypicalPersons;
+import seedu.address.testutil.TypicalPersonsAndAppointments;
 import seedu.address.ui.CommandBox;
 
 /**
@@ -75,7 +75,7 @@ public abstract class AddressBookSystemTest {
      * Returns the data to be loaded into the file in {@link #getDataFileLocation()}.
      */
     protected AddressBook getInitialData() {
-        return TypicalPersons.getTypicalAddressBook();
+        return TypicalPersonsAndAppointments.getTypicalAddressBook();
     }
 
     /**
