+++ src/test/java/systemtests/DeleteBeforeCommandSystemTest.java
@@ -5,7 +5,7 @@ import static seedu.address.logic.commands.DeleteBeforeCommand.MESSAGE_DELETE_PE
 import static seedu.address.testutil.TypicalDates.DATE_FIRST_JAN;
 import static seedu.address.testutil.TypicalDates.INVALID_DATE_DESC;
 import static seedu.address.testutil.TypicalDates.VALID_DATE_DESC;
-import static seedu.address.testutil.TypicalPersons.ALICE;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.ALICE;
 import static seedu.address.testutil.TypicalTags.INVALID_TAG_DESC;
 import static seedu.address.testutil.TypicalTags.TAG_SET_FRIEND;
 import static seedu.address.testutil.TypicalTags.VALID_TAG_DESC_FRIEND;
