+++ src/test/java/systemtests/FindAliasSystemTest.java
@@ -3,14 +3,14 @@ package systemtests;
 import static org.junit.Assert.assertFalse;
 import static seedu.address.commons.core.Messages.MESSAGE_PERSONS_LISTED_OVERVIEW;
 import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
-import static seedu.address.testutil.TypicalPersons.ALICE;
-import static seedu.address.testutil.TypicalPersons.BENSON;
-import static seedu.address.testutil.TypicalPersons.CARL;
-import static seedu.address.testutil.TypicalPersons.DANIEL;
-import static seedu.address.testutil.TypicalPersons.ELLE;
-import static seedu.address.testutil.TypicalPersons.FIONA;
-import static seedu.address.testutil.TypicalPersons.GEORGE;
-import static seedu.address.testutil.TypicalPersons.KEYWORD_MATCHING_MEIER;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.ALICE;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.BENSON;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.CARL;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.DANIEL;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.ELLE;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.FIONA;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.GEORGE;
+import static seedu.address.testutil.TypicalPersonsAndAppointments.KEYWORD_MATCHING_MEIER;
 
 import java.util.ArrayList;
 import java.util.List;
